lambda/requirements.txt
boto3>=1.34
pandas>=2.1
pyarrow>=14.0
google-cloud-bigquery>=3.13
db-dtypes>=1.2
